Change type of baseprojectilespell speed and damage to Stat

+++ include/entity.h
@@ -103,7 +103,7 @@ class Player {
   const BaseProjectileSpell* GetSpell(SpellId id) const;
   void TakeDamage(int damage);
   void TakeHealing(int healing);
-  int CalculateOutgoingDamage(int damage);
+  int CalculateOutgoingDamage(float damage);
   bool IsSpellReady(BaseSpell& spell, float time);
   void AddToInventory(ItemId item_id);
   Inventory inventory_;
